the cache engine configuration item optimization
