+++ cacheLib/src/main/java/com/dvsnier/cache/CacheManager.java
@@ -265,7 +265,6 @@ public class CacheManager extends AbstractCacheManager {
     @Override
     public CacheTransactionSession put(@NonNull String key, Object value, long duration, TimeUnit timeUnit) {
         return getCacheWrapper().getTransaction().put(key, value, duration, timeUnit);
-
     }
 
     @Multiple
