+++ cacheLib/src/main/java/com/dvsnier/cache/base/IEngineInstrument.java
@@ -1,10 +1,11 @@
 package com.dvsnier.cache.base;
 
 import com.dvsnier.cache.config.IAlias;
+import com.dvsnier.cache.config.ICacheAPI;
 
 /**
  * IEngineInstrument
  * Created by dovsnier on 2019-07-24.
  */
-public interface IEngineInstrument extends ICacheEngine, IAlias {
+public interface IEngineInstrument extends ICacheEngine, IAlias, ICacheAPI {
 }
